feat: run connector locally
