--- rust-connectors/sources/http/examp.yaml
@@ -1,11 +1,0 @@
-version: v1
-name: route-data
-type: http-json
-topic: routes-info
-create_topic: true
-parameters:
-  endpoint: https://api.digitransit.fi/routing/v1/routers/hsl/index/graphql
-  method: POST
-  body: '{"query":"{\n  stops {\n    stopId: gtfsId\n    lat\n    lon\n    code\n    stoptimesWithoutPatterns {\n      trip {\n        tripId: gtfsId\n        serviceId\n      }\n      scheduledArrival\n    }\n  }\n}","variables":null}'
-  interval: 3600
-  fluvio-topic: routes-info
