+++ rust-connectors/utils/connector-deploy/Cargo.toml
@@ -2,6 +2,7 @@
 name = "connector-deploy"
 version = "0.1.0"
 edition = "2021"
+description = "Deploy or run locally a connector"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
