+++ rust-connectors/utils/connector-deploy/src/local.rs
@@ -1,0 +1,46 @@
+use std::{io, io::Write, process::Command};
+
+use clap::Parser;
+use fluvio_connectors_common::config::ConnectorConfig;
+
+use crate::build_args;
+
+#[derive(Debug, Parser)]
+pub struct LocalOpt {
+    /// path to the connector config
+    #[clap(short = 'c', long)]
+    config: String,
+}
+
+impl LocalOpt {
+    pub fn execute(self) -> anyhow::Result<()> {
+        let config = ConnectorConfig::from_file(self.config)?;
+
+        let args = build_args(&config);
+        let type_ = &config.type_;
+        let image = format!("infinyon/fluvio-connect-{}:{}", type_, config.version);
+
+        let home = std::env::var("HOME").expect("no home found");
+
+        let mut fluvio_config_path = std::path::PathBuf::from(home);
+        fluvio_config_path.push(".fluvio");
+        fluvio_config_path.push("config");
+        let fluvio_config_mount = format!(
+            "type=bind,source={},target=/home/fluvio/.fluvio/config,readonly",
+            fluvio_config_path.as_os_str().to_string_lossy()
+        );
+
+        let mut child = Command::new("docker")
+            .arg("run")
+            .arg("-i")
+            .arg("--mount")
+            .arg(fluvio_config_mount)
+            .arg(image)
+            .args(args)
+            .spawn()?;
+
+        child.wait().expect("failed while waiting for child");
+
+        Ok(())
+    }
+}
